backend/requirements.txt
wheel
flask
flask-cors
python-dotenv
sqlalchemy==1.4.18
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
email-validator
flask-mail
pyjwt
flask-babel
Flask-Bootstrap
Flask-Moment
requests
python-dateutil
flask-httpauth
empyrical
matplotlib
websocket-client
cryptography
oauthlib
pyOpenSSL
python-binance
mysql-connector-python
Flask-Caching
pymysql
deepdiff
Pillow